content and structure changes
Changed:
- main page - two backgrounds in the main carousel were changed to
Eugene's suggestions
- main page - text on the first slide was changed and now contains the
date and the place
- main page - features title was changed to 'Why it is so different?'
because the texts there are the statements about 'why you should attend
the conference'
- 'about' page - tracks snippet was moved to 'speakers' page
- 'about page' - 'what is wildcard' section was changes to plain design
in order to make it readable

Added:
- 'program' page was added. should be updated when we at least decide
the quantity of sessions
- 'for sponsors' page was added - contain the intro text from the
proposal and the like to the proposal. this info will be removed from
the LaunchRock
- 'speakers' page was added. tracks snippet was moved here in order to
navigate the speakers page.
- 'speakers' page - template for the speaker was made. - 'location' page
- map and the venue address were added
- 'about' page - media package section was added in order to provide
logos for those who want to use them for the conf promo